Final Version 2

+++ src/main/java/com/cg/service/SupervisorService.java
@@ -1,16 +1,16 @@
 package com.cg.service;
 
 import java.util.List;
+
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.cg.entity.Vendor;
-
+import com.cg.exception.HallNotFoundException;
 import com.cg.entity.Hall;
-import com.cg.entity.Supervisor;
 import com.cg.repository.HallRepository;
 import com.cg.repository.SupervisorRepository;
 import com.cg.repository.VendorRepository;
@@ -22,32 +22,52 @@ public class SupervisorService {
 	private VendorRepository vendorRepository;
 
 	@Autowired
-	private SupervisorRepository supervisorRepository;
-
-	@Autowired
 	private HallRepository hallRepository;
 
+	@Autowired
+	private SupervisorRepository supervisorRepository;
+
+	static Logger log = Logger.getLogger(SupervisorService.class.getName());
+
 	public ResponseEntity<Object> addHall(int id, Hall hall) {
+
 		supervisorRepository.getById(id).setHall(hall);
+
+		log.info("Supervisor with ID, " + id + " added hall, with ID, " + hall.getHallId());
+
 		return new ResponseEntity<Object>("Hall added successfully.", HttpStatus.OK);
+
 	}
 
 	public ResponseEntity<Object> removeHall(int id) {
-		supervisorRepository.deleteById(id);
-		;
-		return new ResponseEntity<Object>("Hall deleted successfully.", HttpStatus.OK);
+
+		if (supervisorRepository.existsById(id)) {
+
+			supervisorRepository.deleteById(id);
+			log.info("Supervisor deleted hall, with hall ID, " + id);
+
+			return new ResponseEntity<Object>("Hall deleted successfully.", HttpStatus.OK);
+		}
+
+		log.error("No hall found.");
+
+		throw new HallNotFoundException();
 	}
 
 	public ResponseEntity<Object> getSupervisorHallDetails(int id) {
 
-		Supervisor supervisor = supervisorRepository.getById(id);
+		Hall hall = supervisorRepository.findById(id).get().getHall();
 
-		if (supervisor == null) {
+		if (hall == null) {
 
-			return new ResponseEntity<Object>("Currently, no hall is assigned to supervisor.", HttpStatus.OK);
+			log.error("Supervisor tried to access hall, but no hall assigned.");
 
+			throw new HallNotFoundException();
 		}
-		return new ResponseEntity<Object>(supervisor.getHall(), HttpStatus.OK);
+
+		log.info("Supervisor with ID, " + id + " accessed hall, with ID, " + hall.getHallId());
+
+		return new ResponseEntity<Object>(hall, HttpStatus.OK);
 
 	}
 
@@ -60,10 +80,11 @@ public class SupervisorService {
 		billAmount += hall.getHallPrice();
 
 		List<Vendor> vendors = vendorRepository.findByServices(flower, catering, music, video);
-
 		billAmount += vendors.get(0).getVendorCost();
 
 		billAmount *= 1.18;
+
+		log.info("Hall with ID, " + hallId + ", generated bill: " + billAmount);
 
 		return billAmount;
